+++ src/FastD/Framework/Dispatcher/Handle/AnnotationHandle.php
@@ -63,28 +63,30 @@ class AnnotationHandle extends Dispatch
                 foreach ($methods as $method) {
                     $annotation = $extractor->getMethodAnnotation($method);
                     $routeAnnotation = $extractor->getParameters($annotation, 'Route');
+
                     if (empty($routeAnnotation)) {
                         continue;
                     }
+
+                    $routeMethods = $extractor->getParameters($annotation, 'Methods');
+                    $routeRoles = $extractor->getParameters($annotation, 'Roles');
+                    $routeSchemas = $extractor->getParameters($annotation, 'Schemas');
+                    $routeHosts = $extractor->getParameters($annotation, 'Hosts');
+                    $routeFormats = $extractor->getParameters($annotation, 'Formats');
                     $routeAnnotation['class'] = $className;
                     $routeAnnotation['action'] = $method;
+                    $routeAnnotation['method'] = empty($routeMethods) ? ['ANY'] : json_decode($routeMethods[0], true);
+                    $routeAnnotation['roles'] = isset($routeRoles[0]) ? json_decode($routeRoles[0], true) : [];
+                    $routeAnnotation['hosts'] = isset($routeHosts[0]) ? json_decode($routeHosts[0], true) : [];
+                    $routeAnnotation['schemas'] = isset($routeSchemas[0]) ? json_decode($routeSchemas[0], true) : [];
+                    $routeAnnotation['formats'] = isset($routeFormats[0]) ? json_decode($routeFormats[0], true) : [];
                     $routesAnnotation[] = $routeAnnotation;
                 }
 
                 $routes = function () use ($routesAnnotation) {
                     foreach ($routesAnnotation as $routeAnnotation) {
                         $args = [];
-                        if (!isset($routeAnnotation['method'])) {
-                            $methodName = 'any';
-                        } else {
-                            if (is_array($routeAnnotation['method'])) {
-                                $methodName = 'match';
-                                $args[] = $routeAnnotation['method'];
-                            } else {
-                                $methodName = $routeAnnotation['method'];
-                            }
-                        }
-
+                        $args[] = $routeAnnotation['method'];
                         if (isset($routeAnnotation[0])) {
                             if (!isset($routeAnnotation['name'])) {
                                 $args[] = $routeAnnotation[0];
@@ -97,8 +99,19 @@ class AnnotationHandle extends Dispatch
                         }
 
                         $args[] = $routeAnnotation['class'] . '@' . $routeAnnotation['action'];
+                        $args[] = isset($routeAnnotation['defaults']) ? $routeAnnotation['defaults'] : [];
+                        $args[] = isset($routeAnnotation['requirements']) ? $routeAnnotation['requirements'] : [];
+                        $route = call_user_func_array("\\Routes::match", $args);
 
-                        call_user_func_array("\\Routes::{$methodName}", $args);
+                        if (!empty($routeAnnotation['hosts'])) {
+                            $route->setHost($routeAnnotation['hosts']);
+                        }
+                        if (!empty($routeAnnotation['schemas'])) {
+                            $route->setSchema($routeAnnotation['schemas']);
+                        }
+                        if (!empty($routeAnnotation['formats'])) {
+                            $route->setSchema($routeAnnotation['formats']);
+                        }
                     }
                 };